SQS-379 | processOrderBookActiveOrders
Move OrderbookTestHelper and friends to orderbooktesting package.
